+++ Assets/Scripts/TestDllA/TestDllA_Main.cs
@@ -1,4 +1,4 @@
-﻿#define APPLY_PATCH
+﻿//#define APPLY_PATCH
 
 using System;
 using System.Collections;
